The Perl interpreter cmake version now works under windows, where you do not get dynmically loadable perl modules.
The Perl_cmiss lib is now not darchived and put into perlinterpreter lib but is installed alongside the perlinterpreter library and bought in as a dependency.  This works on both windows and linux.

The debug version of the windows build builds against the debug microsoft c library for consistency with all the other static libraries that cmgui is dependent on.